+++ src/server/HttpTaskServer.java
@@ -1,6 +1,6 @@
 package server;
 
-import server.Handlers.*;
+import server.handlers.*;
 import com.sun.net.httpserver.HttpServer;
 import taskmanager.TaskManager;
 import java.io.IOException;
@@ -16,21 +16,27 @@ public class HttpTaskServer {
     }
 
     public HttpServer createServer() {
+
         try {
             HttpServer httpServer = HttpServer.create(new InetSocketAddress(PORT), 0);
             createServerHandlers(httpServer);
             return httpServer;
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            System.err.println("Произошла ошибка при создании сервера: " + e.getMessage());
+            e.printStackTrace();
+            return null;
         }
+
     }
 
 
     private void createServerHandlers(HttpServer incomingServer) {
+
         incomingServer.createContext("/tasks", new TasksHandler(manager));
         incomingServer.createContext("/subtasks", new SubtasksHandler(manager));
         incomingServer.createContext("/epics", new EpicsHandler(manager));
         incomingServer.createContext("/history", new HistoryHandler(manager));
         incomingServer.createContext("/prioritized", new PrioritizedHandler(manager));
+
     }
 }
