+++ src/server/handlers/adapters/DurationAdapter.java
@@ -1,4 +1,4 @@
-package server.Handlers.Adapters;
+package server.handlers.adapters;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
